+++ src/app/admin/modules/data-management/components/batch-submit/batch-submit.component.spec.ts
@@ -6,7 +6,7 @@ import { HttpErrorResponse } from '@angular/common/http';
 import { Subject } from 'rxjs';
 
 import { BatchSubmitComponent } from './batch-submit.component';
-import { RecordService } from 'src/app/user/modules/records/services/record.service';
+import { RecordService } from 'src/app/shared/services/records/record.service';
 
 
 describe('BatchSubmitComponent', () => {
