+++ src/app/admin/modules/data-management/components/batch-submit/batch-submit.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { RecordService } from 'src/app/user/modules/records/services/record.service';
+import { RecordService } from 'src/app/shared/services/records/record.service';
 import { MatSnackBar } from '@angular/material';
 import { HttpErrorResponse } from '@angular/common/http';
