+++ src/app/admin/modules/program-management/admin-events/components/admin-campus-event-list/admin-campus-event-list.component.spec.ts
@@ -5,8 +5,8 @@ import { Router, ActivatedRoute } from '@angular/router';
 import { of as observableOf, Subject } from 'rxjs';
 import { CampusEvent } from 'src/app/shared/interfaces/event';
 import { Program } from 'src/app/shared/interfaces/program';
-import { ProgramService } from '../../../admin-programs/services/program.service';
-import { EventService } from '../../services/event.service';
+import { ProgramService } from 'src/app/shared/services/programs/program.service';
+import { EventService } from 'src/app/shared/services/events/event.service';
 import { PaginatedResponse } from 'src/app/shared/interfaces/paginated-response';
 import { HAMMER_LOADER } from '@angular/platform-browser';
 import { Location } from '@angular/common';
