+++ src/app/admin/modules/program-management/admin-programs/components/admin-program-form/admin-program-form.component.spec.ts
@@ -16,9 +16,9 @@ import { HttpErrorResponse } from '@angular/common/http';
 import { Program } from 'src/app/shared/interfaces/program';
 import { ProgramForm } from 'src/app/shared/interfaces/program-form';
 import { ProgramCategory } from 'src/app/shared/interfaces/program-category';
-import { ProgramService } from '../../services/program.service';
-import { ProgramFormService} from '../../services/program-form.service';
-import { ProgramCategoryService} from '../../services/program-category.service';
+import { ProgramService } from 'src/app/shared/services/programs/program.service';
+import { ProgramFormService} from 'src/app/shared/services/programs/program-form.service';
+import { ProgramCategoryService} from 'src/app/shared/services/programs/program-category.service';
 import { AUTH_SERVICE } from 'src/app/shared/interfaces/auth-service';
 import { PaginatedResponse } from 'src/app/shared/interfaces/paginated-response';
