+++ src/app/shared/components/off-campus-record-detail/off-campus-record-detail.component.spec.ts
@@ -16,7 +16,7 @@ import {
 import { ActivatedRoute, Router } from '@angular/router';
 
 import { OffCampusRecordDetailComponent } from './off-campus-record-detail.component';
-import { ReviewNoteService } from 'src/app/admin/modules/data-management/services/review-note.service';
+import { ReviewNoteService } from 'src/app/shared/services/records/review-note.service';
 import { ReviewNote } from 'src/app/shared/interfaces/review-note';
 import { Record } from 'src/app/shared/interfaces/record';
 import { Location } from '@angular/common';
