+++ src/app/shared/services/events/event-detail-resolver.service.spec.ts
