+++ src/app/shared/services/programs/program-detail-resolver.service.ts
