+++ src/app/shared/services/records/record-content.service.ts
