+++ src/app/shared/services/records/record-detail-resolver.service.ts
