+++ src/app/user/modules/statistics/components/data-export/data-export.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { Location } from '@angular/common';
-import { RecordService } from 'src/app/user/modules/records/services/record.service';
+import { RecordService } from 'src/app/shared/services/records/record.service';
 import { GenericListComponent } from 'src/app/shared/generics/generic-list/generic-list';
 import { Record } from 'src/app/shared/interfaces/record';
 import { ActivatedRoute, Router } from '@angular/router';
